voting_application/requirements.txt
Adafruit_BBIO==1.2.0
aenum==3.1.11
asgiref==3.5.1
blinker==1.4
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.2.0
cryptography==3.4.7
Cython==0.29.24
dl==0.1.0
docutils==0.17.1
eventlet==0.33.0
Flask_SQLAlchemy==2.5.1
gevent==21.8.0
gpiod==1.5.3
greatfet==2021.2.1
greenlet==1.1.1
hid==1.0.5
HTMLParser==0.0.2
importlib_metadata==4.6.4
inotify==0.2.10
keyring==23.0.1
lockfile==0.12.2
lxml==4.6.3
machine==0.0.1
numpy==1.20.3
ordereddict==1.1
PasteScript==3.2.1
protobuf==3.20.1
pyOpenSSL==22.0.0
python-dotenv==0.20.0
setproctitle==1.2.3
sets==0.3.2
sysv_ipc==1.1.0
toml==0.10.2
tornado==6.1
typing_extensions==3.10.0.0
utime==1.0.3
watchdog==1.0.2
xmlrpclib==1.0.1
zipp==3.5.0
